+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Conway/Cert.hs
@@ -13,6 +13,7 @@
 
 module Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Cert () where
 
+import Cardano.Ledger.Address (RewardAccount)
 import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.CertState
 import Cardano.Ledger.Coin
@@ -20,13 +21,14 @@ import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Governance
 import Cardano.Ledger.Conway.Rules
 import Cardano.Ledger.Conway.TxCert
-import Cardano.Ledger.Credential
 import Cardano.Ledger.EpochBoundary
-import Cardano.Ledger.Keys
 import Cardano.Ledger.Shelley.LedgerState
-import Cardano.Ledger.UMap (dRepMap, rewardMap, sPoolMap)
+import Cardano.Ledger.UMap (depositMap)
+import Data.Bifunctor (Bifunctor (..))
 import Data.Functor.Identity (Identity)
 import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as Map
+import qualified Data.OMap.Strict as OMap
 import qualified Data.VMap as VMap
 import Lens.Micro
 import qualified Lib as Agda
@@ -41,30 +43,30 @@ instance
   ( SpecTranslate ctx (PParamsHKD Identity era)
   , SpecRep (PParamsHKD Identity era) ~ Agda.PParams
   , Inject ctx (VotingProcedures era)
-  , Inject ctx (Map (Network, Credential 'Staking (EraCrypto era)) Coin)
+  , Inject ctx (Map (RewardAccount (EraCrypto era)) Coin)
   ) =>
   SpecTranslate ctx (CertEnv era)
   where
-  type SpecRep (CertEnv era) = Agda.CertEnv'
+  type SpecRep (CertEnv era) = Agda.CertEnv
   toSpecRep CertEnv {..} = do
     votes <- askCtx @(VotingProcedures era)
-    withdrawals <- askCtx @(Map (Network, Credential 'Staking (EraCrypto era)) Coin)
-    Agda.MkCertEnv'
+    withdrawals <- askCtx @(Map (RewardAccount (EraCrypto era)) Coin)
+    Agda.MkCertEnv
       <$> toSpecRep ceCurrentEpoch
       <*> toSpecRep cePParams
       <*> toSpecRep votes
       <*> toSpecRep withdrawals
 
 instance SpecTranslate ctx (CertState era) where
-  type SpecRep (CertState era) = Agda.CertState'
+  type SpecRep (CertState era) = Agda.CertState
   toSpecRep CertState {..} =
-    Agda.MkCertState'
+    Agda.MkCertState
       <$> toSpecRep certDState
       <*> toSpecRep certPState
       <*> toSpecRep certVState
 
-instance SpecTranslate ctx (ConwayTxCert era) where
-  type SpecRep (ConwayTxCert era) = Agda.TxCert
+instance Era era => SpecTranslate ctx (ConwayTxCert era) where
+  type SpecRep (ConwayTxCert era) = Agda.DCert
 
   toSpecRep (ConwayTxCertPool p) = toSpecRep p
   toSpecRep (ConwayTxCertGov c) = toSpecRep c
@@ -76,31 +78,26 @@ instance
   , SpecTranslate ctx (PParamsHKD StrictMaybe era)
   , SpecRep (PParamsHKD StrictMaybe era) ~ Agda.PParamsUpdate
   , SpecRep (TxOut era) ~ Agda.TxOut
-  , SpecTranslate ctx (TxOut era)
+  , SpecTranslate (Map (DepositPurpose (EraCrypto era)) Coin) (TxOut era)
   ) =>
   SpecTranslate ctx (LedgerState era)
   where
-  type SpecRep (LedgerState era) = Agda.LedgerState
+  type SpecRep (LedgerState era) = Agda.LState
 
-  toSpecRep (LedgerState {..}) =
-    Agda.MkLedgerState
-      <$> toSpecRep lsUTxOState
-      <*> toSpecRep (utxosGovState lsUTxOState ^. proposalsGovStateL)
-      <*> agdaCertState lsCertState
-    where
-      agdaCertState (CertState (VState {..}) pState (DState {..})) =
-        Agda.MkCertState
-          <$> ( Agda.MkDState
-                  <$> toSpecRep (dRepMap dsUnified)
-                  <*> toSpecRep (sPoolMap dsUnified)
-                  <*> toSpecRep (rewardMap dsUnified)
-              )
-          <*> toSpecRep pState
-          <*> ( Agda.MkGState
-                  <$> toSpecRep (drepExpiry <$> vsDReps)
-                  <*> toSpecRep
-                    (committeeCredentialToStrictMaybe <$> csCommitteeCreds vsCommitteeState)
-              )
+  toSpecRep (LedgerState {..}) = do
+    let
+      props = utxosGovState lsUTxOState ^. proposalsGovStateL
+      deposits =
+        Map.unions
+          [ Map.mapKeys CredentialDeposit $ depositMap (lsCertState ^. certDStateL . dsUnifiedL)
+          , Map.mapKeys PoolDeposit $ lsCertState ^. certPStateL . psDepositsL
+          , fmap drepDeposit . Map.mapKeys DRepDeposit $ lsCertState ^. certVStateL . vsDRepsL
+          , Map.fromList . fmap (bimap GovActionDeposit gasDeposit) $ OMap.assocList (props ^. pPropsL)
+          ]
+    Agda.MkLState
+      <$> withCtx deposits (toSpecRep lsUTxOState)
+      <*> toSpecRep props
+      <*> withCtx deposits (toSpecRep lsCertState)
 
 instance
   ( EraPParams era
@@ -112,7 +109,7 @@ instance
   , Inject ctx [GovActionState era]
   , ToExpr (PParamsHKD StrictMaybe era)
   , SpecRep (TxOut era) ~ Agda.TxOut
-  , SpecTranslate ctx (TxOut era)
+  , SpecTranslate (Map (DepositPurpose (EraCrypto era)) Coin) (TxOut era)
   ) =>
   SpecTranslate ctx (EpochState era)
   where
@@ -170,7 +167,7 @@ instance
   , Inject ctx [GovActionState era]
   , ToExpr (PParamsHKD StrictMaybe era)
   , SpecRep (TxOut era) ~ Agda.TxOut
-  , SpecTranslate ctx (TxOut era)
+  , SpecTranslate (Map (DepositPurpose (EraCrypto era)) Coin) (TxOut era)
   ) =>
   SpecTranslate ctx (NewEpochState era)
   where