+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Conway/Deleg.hs
@@ -39,15 +39,15 @@ instance
   ) =>
   SpecTranslate ctx (ConwayDelegEnv era)
   where
-  type SpecRep (ConwayDelegEnv era) = Agda.DelegEnv'
+  type SpecRep (ConwayDelegEnv era) = Agda.DelegEnv
 
   toSpecRep ConwayDelegEnv {..} =
-    Agda.MkDelegEnv'
+    Agda.MkDelegEnv
       <$> toSpecRep cdePParams
       <*> toSpecRep (Map.mapKeys (hashToInteger . unKeyHash) cdePools)
 
 instance SpecTranslate ctx (ConwayDelegCert c) where
-  type SpecRep (ConwayDelegCert c) = Agda.TxCert
+  type SpecRep (ConwayDelegCert c) = Agda.DCert
 
   toSpecRep (ConwayRegCert c d) =
     Agda.Delegate
@@ -78,10 +78,10 @@ instance SpecTranslate ctx (ConwayDelegPredFailure era) where
   toSpecRep e = pure . OpaqueErrorString . show $ toExpr e
 
 instance SpecTranslate ctx (DState era) where
-  type SpecRep (DState era) = Agda.DState'
+  type SpecRep (DState era) = Agda.DState
 
   toSpecRep DState {..} =
-    Agda.MkDState'
+    Agda.MkDState
       <$> toSpecRep (UMap.dRepMap dsUnified)
       <*> toSpecRep (UMap.sPoolMap dsUnified)
       <*> toSpecRep (UMap.rewardMap dsUnified)